% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca-exchange-annotate-maf.R
\name{brca_exchange_annotate_maf}
\alias{brca_exchange_annotate_maf}
\alias{query_brca_exchange}
\alias{brca_annotate_maf}
\title{Annotate BRCA1/2 variants}
\source{
\url{https://brcaexchange.org}
}
\usage{
query_brca_exchange(gene, start, end, ref, alt, use_api = FALSE)

brca_annotate_maf(maf, use_api = FALSE)
}
\arguments{
\item{gene}{Query gene.}

\item{start}{Variant start position, in hg19.}

\item{end}{Variant end position, in hg19.}

\item{ref}{Reference allele.}

\item{alt}{Alternate allele.}

\item{use_api}{If \code{TRUE} uses BRCA exchange API, otherwise static table.}

\item{maf}{Input MAF.}
}
\value{
Annotated MAF with columns \code{brca_exchange_enigma}, \code{brca_exchange_clinvar} variant annotation from ENIGMA and ClinVar, respectively, and \code{brca_exchange_id} indicating database variant ID.
}
\description{
Adds functional BRCA1 and BRCA2 variant annotation, using BRCA Exchange API (see URL). Annotation comes from ENIGMA and ClinVar.
Not that variant alleles in a MAF are in the column \code{Tumor_Seq_Allele2} by default.
}
\examples{
\donttest{query_brca_exchange('BRCA1', 41276045, 41276046, 'CT', '-')}
  
}
