% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coding_mutations}
\alias{coding_mutations}
\title{Variant classification of coding mutations}
\format{List of variants considered coding, by VEP variant class labels.}
\usage{
coding_mutations
}
\description{
Variant classification of coding mutations
}
\keyword{datasets}
