% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot-annotate-maf.R
\name{hotspot_annotate_maf}
\alias{hotspot_annotate_maf}
\title{Annotate somatic hotspot mutations}
\usage{
hotspot_annotate_maf(maf, hotspot_tbl = NULL)
}
\arguments{
\item{maf}{Input MAF.}

\item{hotspot_tbl}{Custom table of hotspots.}
}
\value{
Annotated MAF with columns \code{snv_hotspot}, \code{threeD_hotspot}, \code{indel_hotspot_type} and \code{Hotspot} indicating types of hotspots. Note that the \code{Hotspot} column does not includes 3D hotspots.
}
\description{
Adds hotspot annotation to VEP-annotated MAF. Sources of default hotspots below.
}
