% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/hotspots.R
\docType{data}
\name{hotspots}
\alias{hotspots}
\title{Mutation hotspots}
\format{data.frame with SNV and indel hotspots from Taylor lab.}
\source{
\url{http://cancerhotspots.org}

\url{https://www.ncbi.nlm.nih.gov/pubmed/26619011}

\url{https://www.ncbi.nlm.nih.gov/pubmed/29247016}

\url{https://www.ncbi.nlm.nih.gov/pubmed/28115009}
}
\usage{
hotspots

hotspots
}
\description{
Mutation hotspots

Hotspots
}
\keyword{datasets}
