% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truncating_mutations}
\alias{truncating_mutations}
\title{Variant classification of truncating mutations}
\format{List of variants considered protein truncating, by VEP variant class labels.}
\usage{
truncating_mutations
}
\description{
Variant classification of truncating mutations
}
\keyword{datasets}
