% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca-exchange-variants.R
\docType{data}
\name{brca_exchange_variants}
\alias{brca_exchange_variants}
\title{BRCA exchange variants}
\format{Table of annotated BRCA1/2 variants in BRCA exchange database.}
\source{
\url{http://brcaexchange.org}
}
\usage{
brca_exchange_variants
}
\description{
BRCA exchange variants
}
\keyword{datasets}
